refactor : TechStackErrorMessage 분리 및 정의
