+++ src/main/java/sixgaezzang/sidepeek/common/exception/message/TechStackErrorMessage.java
@@ -1,0 +1,23 @@
+package sixgaezzang.sidepeek.common.exception.message;
+
+import static sixgaezzang.sidepeek.common.util.CommonConstant.MAX_CATEGORY_LENGTH;
+import static sixgaezzang.sidepeek.common.util.CommonConstant.MAX_TECH_STACK_COUNT;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public class TechStackErrorMessage {
+
+    // Tech Stack List
+    public static final String TECH_STACKS_IS_NULL = "기술 스택들을 입력해주세요.";
+    public static final String TECH_STACKS_OVER_MAX_COUNT =
+        "기술 스택은 " + MAX_TECH_STACK_COUNT + "개 미만이어야 합니다.";
+    public static final String TECH_STACK_IS_DUPLICATED = "같은 카테고리 내에 같은 기술 스택이 있습니다.";
+
+    // Tech Stack
+    public static final String CATEGORY_IS_NULL = "기술 스택 카테고리를 입력해주세요.";
+    public static final String CATEGORY_OVER_MAX_LENGTH =
+        "기술 스택 카테고리는 " + MAX_CATEGORY_LENGTH + "자 이하여야 합니다.";
+
+}
